Changed when to emit 'logSenderTotalMessagesRead'
The logSenderTotalMessagesRead metric is only emitted when a component
actively calls SendAppLog or SendErrorLog.

The 'logSenderTotalMessagesRead' metric type has also been changed from
a ValueMetric to a CounterEvent. The metric was a counter and it did not
make sense to emit a ValueMetric with unit 'count'.

The LogSender type now uses the metrics package to send the
logSenderTotalMessagesRead metric, instead of spawing its own goroutine.
Due to this change it's no longer required to pass in a
'counterEmitDuration' when creating a LogSender.

[#103444504]