Replace exported interface types with exported concrete types
The concrete types were documented pretty well, but none of the documentation shows up on
godoc because the concrete types weren't exported.  Since we're trying to move away from using
interface types as a source of truth, we removed most of those source-of-truth interface
types.

While we were at it, we found and fixed a flaky test and cleaned up some old, crusty docs.

[#118508897]

Signed-off-by: Jason Keene <jkeene@pivotal.io>